This fixes the implementation of the key generation in managed mode
Instead of using the passphrase to creat the partial key of the health authority
for the QRcodes, this choses a random key.
The organizer will have to store the ctxtha and mpk public keys and use them in
case of a notification.
This allows the organizer to be protected against coercion attacks: as the organizer
doesn't know the private health authority partial key, she cannot be coerced
into sending fake notification keys.
