+++ .gitattributes
@@ -1,0 +1,1 @@
+package-lock.json -diff
