+++ app/index.ts
@@ -1,7 +1,8 @@
 import * as v1 from './v1';
 import * as v1_1 from './v1_1';
 import {HealthAuthority, Location, Visit,
-  Organizer, Room} from './v2';
+  Room} from './v2';
+import {Organizer} from './v2_1';
 
 import {mcl, sodium, waitReady} from '@c4dt/libcrowdnotifier';
 import * as lib from '@c4dt/libcrowdnotifier';
