+++ app/v2_1/index.ts
@@ -1,0 +1,3 @@
+import {Organizer, Room} from './managed';
+
+export {Organizer, Room};
