+++ lib/src/v2/crowd_notifier_primitives.ts
@@ -2,7 +2,7 @@ import {
   compare,
   crypto_box_keypair,
   crypto_box_seal,
-  crypto_box_seal_open, from_string,
+  crypto_box_seal_open,
   IKeyPair,
   randombytes_buf,
 } from 'libsodium-wrappers-sumo';
@@ -10,10 +10,10 @@ import {dec, enc, IEncryptedData,
   keyDer, keyGen, NONCE_LENGTH} from './ibe_primitives';
 import mcl from 'mcl-wasm';
 import {
-  IEntryProof, ILocationData, IMasterTrace, IOrganizerData,
+  IEntryProof, ILocationData, IMasterTrace,
   IPreTrace, ITrace, ITraceProof,
 } from './structs';
-import {baseG2, genId} from './helpers';
+import {genId} from './helpers';
 
 
 /**
@@ -26,50 +26,6 @@ import {baseG2, genId} from './helpers';
  */
 export function setupHA(): IKeyPair {
   return crypto_box_keypair();
-}
-
-/**
- * Generates the static part for v2.1 from a pass phrase. The idea is to
- * have a static part that doesn't change, to secure the notifications without
- * having to keep too many keys.
- * @param pkh public key of the health authority
- * @param pp pass phrase, if possible 256 bits entropy or more
- */
-export function genOrgStatic(pkh: Uint8Array, pp: string): IOrganizerData {
-  const mskO = new mcl.Fr();
-  mskO.setHashOf(from_string(pp));
-  const mpkO = mcl.mul(baseG2(), mskO);
-
-  const mskha = new mcl.Fr();
-  mskha.setHashOf(from_string(`healthAuthority:${pp}`));
-  const mpkha = mcl.mul(baseG2(), mskha);
-  const mpk = mcl.add(mpkO, mpkha);
-
-  const ctxtha = crypto_box_seal(mskha.serialize(), pkh);
-
-  return {mpk, mskO, ctxtha};
-}
-
-/**
- * Generates the code for v2.1 with an organization master key.
- * @param org from the genOrgStatic
- * @param info about the room
- * @return the QRCodeContent to be printed in QR codes
- */
-export function genOrgCode(org: IOrganizerData, info: Uint8Array):
-    ILocationData {
-  const nonce1 = randombytes_buf(NONCE_LENGTH);
-  const nonce2 = randombytes_buf(NONCE_LENGTH);
-
-  const pEnt = {nonce1, nonce2};
-  const mtr = {
-    mpk: org.mpk,
-    mskl: org.mskO,
-    ctxtha: org.ctxtha,
-    info, nonce1, nonce2,
-  };
-
-  return {ent: org.mpk, pEnt, mtr};
 }
 
 /**
